moving file watcher thread test into 'BundleCachingTest'
