+++ brjs-core-tests/src/test-integration/java/org/bladerunnerjs/spec/brjs/appserver/AppServerTest.java
@@ -345,19 +345,4 @@ public class AppServerTest extends SpecTest
 		then(appServer).requestForUrlContains("/app1/v/dev/no-such-content-plugin", "Error 404");
 	}
 	
-	@Test
-	public void fileWatcherWatchesFolderBrjsAppsAtTheSameLevelAsSdk() throws Exception {
-		given(brjs).hasBeenAuthenticallyCreatedWithFileWatcherThread();
-			app1 = brjs.app("app1");
-			aspect = app1.defaultAspect();
-			given(app1).hasBeenCreated()
-			.and(aspect).hasBeenCreated()
-			.and(aspect).indexPageHasContent("require('appns/App')")
-			.and(aspect).classFileHasContent("App", "// App.js")
-			.and(aspect).hasReceivedRequest("js/dev/combined/bundle.js");
-		when(aspect).indexPageRefersToWithoutNotifyingFileRegistry("require('appns/AppClass')")
-			.and(aspect).fileHasContentsWithoutNotifyingFileRegistry("src/AppClass.js", "// AppClass.js");
-		then(aspect).devResponseEventuallyContains("js/dev/combined/bundle.js", "AppClass.js", response)
-			.and(response).doesNotContainText("App.js");
-	}
 }
